+++ database/migrations/2018_09_29_172108_create_contact_category_table.php
@@ -14,10 +14,10 @@ class CreateContactCategoryTable extends Migration
     public function up()
     {
         Schema::create('contact_category', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('category_id')->unsigned()->default(0);
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('category_id')->unsigned()->default(0);
             $table->foreign('category_id')->references('id')->on('categories');
-            $table->integer('contact_id')->unsigned()->default(0);
+            $table->unsignedBigInteger('contact_id')->unsigned()->default(0);
             $table->foreign('contact_id')->references('id')->on('contacts');
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
