+++ database/migrations/2019_08_24_121202_create_seasons_goods_table.php
@@ -14,10 +14,10 @@ class CreateSeasonsGoodsTable extends Migration
     public function up()
     {
         Schema::create('seasons_goods', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('good_id')->unsigned()->default(0);
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('good_id')->unsigned()->default(0);
             $table->foreign('good_id')->references('id')->on('goods')->onDelete('cascade');;
-            $table->integer('season_id')->unsigned()->default(0);
+            $table->unsignedBigInteger('season_id')->unsigned()->default(0);
             $table->foreign('season_id')->references('id')->on('seasons')->onDelete('cascade');;
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
