+++ database/migrations/2019_08_24_155524_create_fabrics_goods_table.php
@@ -14,10 +14,10 @@ class CreateFabricsGoodsTable extends Migration
     public function up()
     {
         Schema::create('fabrics_goods', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('good_id')->unsigned()->default(0);
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('good_id')->unsigned()->default(0);
             $table->foreign('good_id')->references('id')->on('goods')->onDelete('cascade');;
-            $table->integer('fabric_id')->unsigned()->default(0);
+            $table->unsignedBigInteger('fabric_id')->unsigned()->default(0);
             $table->foreign('fabric_id')->references('id')->on('fabrics')->onDelete('cascade');;
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
