+++ database/migrations/2019_08_24_162105_create_decorations_goods_table.php
@@ -14,10 +14,10 @@ class CreateDecorationsGoodsTable extends Migration
     public function up()
     {
         Schema::create('decorations_goods', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('good_id')->unsigned()->default(0);
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('good_id')->unsigned()->default(0);
             $table->foreign('good_id')->references('id')->on('goods')->onDelete('cascade');;
-            $table->integer('decoration_id')->unsigned()->default(0);
+            $table->unsignedBigInteger('decoration_id')->unsigned()->default(0);
             $table->foreign('decoration_id')->references('id')->on('decorations')->onDelete('cascade');;
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
