+++ database/migrations/2019_08_24_162301_create_producttypes_goods_table.php
@@ -14,10 +14,10 @@ class CreateProducttypesGoodsTable extends Migration
     public function up()
     {
         Schema::create('producttypes_goods', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('good_id')->unsigned()->default(0);
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('good_id')->unsigned()->default(0);
             $table->foreign('good_id')->references('id')->on('goods')->onDelete('cascade');;
-            $table->integer('producttype_id')->unsigned()->default(0);
+            $table->unsignedBigInteger('producttype_id')->unsigned()->default(0);
             $table->foreign('producttype_id')->references('id')->on('product_types')->onDelete('cascade');;
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
