+++ database/migrations/2019_10_27_102031_create_merchants_sites_menu_table.php
@@ -14,12 +14,12 @@ class CreateMerchantsSitesMenuTable extends Migration
     public function up()
     {
         Schema::create('merchants_sites_menu', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('site_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('site_id')->unsigned();
             $table->foreign('site_id')->references('id')->on('sites')->onDelete('cascade');
             $table->string('menu_name');
             $table->string('slug');
-            $table->integer('page_type_id')->unsigned();
+            $table->unsignedBigInteger('page_type_id')->unsigned();
             $table->foreign('page_type_id')->references('id')->on('page_types')->onDelete('cascade');
             $table->timestamps();
         });
